+++ cmd/kubectl-kubtest/commands/scripts/common.go
@@ -1,8 +1,8 @@
 package scripts
 
 import (
-	"github.com/kubeshop/kubtest/pkg/api/client"
-	"github.com/kubeshop/kubtest/pkg/api/kubtest"
+	"github.com/kubeshop/kubtest/pkg/api/v1/client"
+	"github.com/kubeshop/kubtest/pkg/api/v1/kubtest"
 	"github.com/kubeshop/kubtest/pkg/ui"
 	"github.com/spf13/cobra"
 )
