+++ cmd/kubectl-kubtest/commands/scripts/listrenderer.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"text/template"
 
-	"github.com/kubeshop/kubtest/pkg/api/kubtest"
+	"github.com/kubeshop/kubtest/pkg/api/v1/kubtest"
 	"github.com/kubeshop/kubtest/pkg/ui"
 	"github.com/spf13/cobra"
 )
