+++ examples/postman/Kubeshop-failed.postman_collection.json
@@ -1,0 +1,41 @@
+{
+	"info": {
+		"_postman_id": "026a25d6-f311-4561-8ce0-64a824edf7b1",
+		"name": "Kubeshop Failed",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+	},
+	"item": [
+		{
+			"name": "Not Existing",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"pm.test(\"Status code is 200\", function () {",
+							"    pm.response.to.have.status(200);",
+							"});"
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://kubeshop.io/non-existing",
+					"protocol": "https",
+					"host": [
+						"kubeshop",
+						"io"
+					],
+					"path": [
+						"non-existing"
+					]
+				}
+			},
+			"response": []
+		}
+	]
+}
