+++ examples/postman/Kubeshop.postman_collection.json
@@ -1,0 +1,77 @@
+{
+	"info": {
+		"_postman_id": "8af42c21-3e31-49c1-8b27-d6e60623a180",
+		"name": "Kubeshop",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+	},
+	"item": [
+		{
+			"name": "Home",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"pm.test(\"Body matches string\", function () {",
+							"    pm.expect(pm.response.text()).to.include(\"K8s Accelerator\");",
+							"});"
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://kubeshop.io/",
+					"protocol": "https",
+					"host": [
+						"kubeshop",
+						"io"
+					],
+					"path": [
+						""
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Team",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"pm.test(\"Status code is 200\", function () {",
+							"    pm.response.to.have.status(200);",
+							"});",
+							"",
+							"pm.test(\"Body matches string\", function () {",
+							"    pm.expect(pm.response.text()).to.include(\"Jacek Wysocki\");",
+							"});"
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://kubeshop.io/our-team",
+					"protocol": "https",
+					"host": [
+						"kubeshop",
+						"io"
+					],
+					"path": [
+						"our-team"
+					]
+				}
+			},
+			"response": []
+		}
+	]
+}
