+++ examples/postman/KubtestHealthcheck.postman_collection.json
@@ -1,0 +1,39 @@
+{
+	"info": {
+		"_postman_id": "3d9a6be2-bd3e-4cf7-89ca-354103aab4a7",
+		"name": "KubTest",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+	},
+	"item": [
+		{
+			"name": "Health",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"pm.test(\"Status code is 200\", function () {",
+							"    pm.response.to.have.status(200);",
+							"});"
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "{{URI}}/health",
+					"host": [
+						"{{URI}}"
+					],
+					"path": [
+						"health"
+					]
+				}
+			},
+			"response": []
+		}
+	]
+}
