+++ examples/postman/LongRun.postman_collection.json
@@ -1,0 +1,169 @@
+{
+	"info": {
+		"_postman_id": "543efea0-2137-4289-ae72-55b8fd121c88",
+		"name": "LongRun",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+	},
+	"item": [
+		{
+			"name": "kubeshop homepage",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							""
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://kubeshop.io",
+					"protocol": "https",
+					"host": [
+						"kubeshop",
+						"io"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "kubeshop team",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"// save team members to variable",
+							"const $ = cheerio.load(pm.response.text());",
+							"let team = [];",
+							"$('.bio-title').each(function(i,e) {",
+							"   team.push($(e).text());",
+							"});",
+							"",
+							"pm.environment.set(\"team\", team);"
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://kubeshop.io/our-team",
+					"protocol": "https",
+					"host": [
+						"kubeshop",
+						"io"
+					],
+					"path": [
+						"our-team"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Save team to requestBin",
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://en1gyetbc4hqz0h.m.pipedream.net?team={{team}}",
+					"protocol": "https",
+					"host": [
+						"en1gyetbc4hqz0h",
+						"m",
+						"pipedream",
+						"net"
+					],
+					"query": [
+						{
+							"key": "team",
+							"value": "{{team}}"
+						}
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Get all docs pages",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"const $ = cheerio.load(pm.response.text());",
+							"",
+							"const links = []",
+							"$('.md-nav__link').each((i,e) => links.push($(e).attr('href')));",
+							"console.log(links);",
+							"",
+							"for(i in links) {",
+							"    pm.sendRequest('https://kubeshop.github.io/kubtest/'+links[i], function (err, response) {",
+							"        setTimeout(function(){ ",
+							"            console.log($('h1', response.text()).text());",
+							"        }, 31000);",
+							"    });",
+							"",
+							"}",
+							""
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://kubeshop.github.io/kubtest/",
+					"protocol": "https",
+					"host": [
+						"kubeshop",
+						"github",
+						"io"
+					],
+					"path": [
+						"kubtest",
+						""
+					]
+				}
+			},
+			"response": []
+		}
+	],
+	"event": [
+		{
+			"listen": "prerequest",
+			"script": {
+				"type": "text/javascript",
+				"exec": [
+					""
+				]
+			}
+		},
+		{
+			"listen": "test",
+			"script": {
+				"type": "text/javascript",
+				"exec": [
+					""
+				]
+			}
+		}
+	],
+	"variable": [
+		{
+			"key": "JACEK_POSITION",
+			"value": "none"
+		}
+	]
+}
