+++ examples/postman/simple.postman_collection.json
@@ -1,0 +1,67 @@
+{
+	"info": {
+		"_postman_id": "3d9a6be2-bd3e-4cf7-89ca-354103aab4a7",
+		"name": "KubeTest",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+	},
+	"item": [
+		{
+			"name": "Google",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"    pm.test(\"Successful GET request\", function () {",
+							"        pm.expect(pm.response.code).to.be.oneOf([200, 201, 202]);",
+							"    });"
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://google.com",
+					"protocol": "https",
+					"host": [
+						"google",
+						"com"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Kasia.in Homepage",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"exec": [
+							"pm.test(\"Body matches string\", function () {",
+							"    pm.expect(pm.response.text()).to.include(\"PRZEPIS NA CHLEB\");",
+							"});"
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://kasia.in",
+					"protocol": "https",
+					"host": [
+						"kasia",
+						"in"
+					]
+				}
+			},
+			"response": []
+		}
+	]
+}
