+++ pkg/api/v1/client/proxy_test.go
