+++ pkg/api/v1/kubtest/model_execution_result_extended.go
