+++ pkg/api/v1/kubtest/model_script_extended.go
