+++ pkg/mapper/executions/sumary_test.go
@@ -3,7 +3,7 @@ package executions
 import (
 	"testing"
 
-	"github.com/kubeshop/kubtest/pkg/api/kubtest"
+	"github.com/kubeshop/kubtest/pkg/api/v1/kubtest"
 	"github.com/stretchr/testify/assert"
 )
