+++ pkg/mapper/scripts/kube_api.go
@@ -2,7 +2,7 @@ package scripts
 
 import (
 	scriptsV1 "github.com/kubeshop/kubtest-operator/apis/script/v1"
-	"github.com/kubeshop/kubtest/pkg/api/kubtest"
+	"github.com/kubeshop/kubtest/pkg/api/v1/kubtest"
 )
 
 func MapScriptListKubeToAPI(crScripts scriptsV1.ScriptList) (scripts []kubtest.Script) {
